Implement ‘-’, ‘+’, and ‘enter’ keys
implemented as `Motions.MoveToFirstCharacterOfLineUp` and `Motions.MoveToFirstCharacterOfLineDown`

I simply compose either of the functions `MoveUp` or `MoveDown` with `MoveToFirstCharacterOfLine`.

Includes links to the Vim documentation for these motions in the docs folder

The enter key was already mapped to `Motions.MoveDown`, but its functionality is actually more complicated than that.