Use ::setInputEnabled to disable input when available
The approach of preempting textinput events to prevent handling of input
outside of insert mode doesn't mesh well with the new React editor
because we don't use jQuery to process input events. I added an explicit
API for disabling input in atom/atom@64870c733, which is used here if
available. We retain the old interception approach for the time being
for compatibility with the existing editor.