:art: simplify return value of select() method for motion and text-object
return true or false, unless all selection.isEmpty(), it return true.
Before this commit return value was array of boolean, I believe its not
need to be array, since currently number of selection is not vary the
behavior, and number of selection may change by mergeSelection().
So number of selection is not important in this context.