:beer: Noticed, InsertAtOccurrence is totally unnecessary
Since we uniformly treat it as target.
Don’t care if its is occur or not, just treat it as mutation TARGET.

And removed, old aliased InsertAtStartOfSelection.
and InsertAtEndOfSelection.

Update keymap in has-occurrence scope from i and a to I and A.
Visual-Mode’s I, A is already defined since its just InsertAtTarget
regardless of occurrence is not.