:fire: meaningless spec
Failed in atom 1.11.1.
Comment out newPane.addItem(editor) line test pass.
But this is not related to vmp and this spec is for vim-mode which
define command on each text-editor instances but vmp isn’t, vmp just
define commands to scope-based only once. So no risk to duplicate
command registration.
Thus remove!