Allow numpad0 to be used when entering a number before a command
Discovered that this wasn't working and then used the keybinding debugger (ctrl+.) to figure out the issue. Checked whether this works by adding this line to my local keymap. It fixes the problem.

Without this, typing something like `420G` fails while typing `0` because `numpad0` isn't mapped correctly. The `numpad0` key always takes you to the front of the line without this change.