Fix some keyboard shortcuts for layouts with AltGr
In many keyboard layouts, such as German, French and Swedish, some
characters have to be typed using the AltGr key. If such a key is the
non-first key of a VMP keyboard shortcut _sequence,_ the sequence fails
to trigger. This is because Atom receive an 'altgraph' key press
inbetween. For example, the shortcut `g ~` is seen as `g altgraph ~`.
'altgraph' seems to be handled just like any other key, so just like
`g a ~` does not exist, neither does `g altgraph ~`.

This commit duplicates shortcuts containing problematic AltGr keys. For
example, in addition to the `g ~` shortcut I've added `g altgraph ~`.

The keys I did this for are: `~ $ | [ ] { }`.

Fixes #661.